Added test runs for pooled DBs.
Moved testing logic from config_generator to test runner.
Fixed cleaning up code on MySQL.
Test runs no re-create connections for each test, as we have a new event loop, and weird stuff happened.

The issue with pooled transactions is that we are guaranteed to get a different connection than the one that is in a transaction for operations. Should use ContextVar to get connection in case it exists instead of the pool?